Split login and registration tests into their own modules
